taxメソッドの戻り値が整数となるよう修正

+++ src/Review01.java
@@ -5,13 +5,13 @@ public class Review01 {
         //p：税抜き価格、ｒ：税率、result：税額、result2：税込み価格
         int p=2562;
         double r=0.1;
-        double result=tax(p,r);
-        int result2=p+(int)result;
-        System.out.println(p+"円の商品の税込価格は"+result2+"円（消費税は"+(int)result+"円）です。");
+        int result=tax(p,r);
+        int result2=p+result;
+        System.out.println(p+"円の商品の税込価格は"+result2+"円（消費税は"+result+"円）です。");
 
     }
-    public static double tax(int p,double r){
-        double result=p*r;
+    public static int tax(int p,double r){
+        int result=(int)(p*r);
         return result;
     }
